requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.12
django-heroku==0.3.1
django-import-export==2.8.0
django-jet==1.0.8
django-js-asset==2.0.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gitdb==4.0.9
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.5.0
python-dotenv==0.19.2
python-monkey-business==1.0.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
tablib==3.2.0
typing-extensions==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.14.0
xlrd==2.0.1
xlwt==1.3.0
